抽离resource相关成独立组件
