+++ composer.json
@@ -123,6 +123,7 @@
         "hyperf/rate-limit": "self.version",
         "hyperf/redis": "self.version",
         "hyperf/retry": "self.version",
+        "hyperf/resource": "self.version",
         "hyperf/server": "self.version",
         "hyperf/service-governance": "self.version",
         "hyperf/session": "self.version",
@@ -200,6 +201,7 @@
             "Hyperf\\RateLimit\\": "src/rate-limit/src/",
             "Hyperf\\Redis\\": "src/redis/src/",
             "Hyperf\\Retry\\": "src/retry/src/",
+            "Hyperf\\Resource\\": "src/resource/src/",
             "Hyperf\\RpcClient\\": "src/rpc-client/src/",
             "Hyperf\\RpcServer\\": "src/rpc-server/src/",
             "Hyperf\\Rpc\\": "src/rpc/src/",
@@ -271,6 +273,7 @@
             "HyperfTest\\RateLimit\\": "src/rate-limit/tests/",
             "HyperfTest\\Redis\\": "src/redis/tests/",
             "HyperfTest\\Retry\\": "src/retry/tests/",
+            "HyperfTest\\Resource\\": "src/resource/tests/",
             "HyperfTest\\Rpc\\": "src/rpc/tests/",
             "HyperfTest\\Server\\": "src/server/tests/",
             "HyperfTest\\ServiceGovernance\\": "src/service-governance/tests/",
@@ -337,6 +340,7 @@
                 "Hyperf\\RateLimit\\ConfigProvider",
                 "Hyperf\\Redis\\ConfigProvider",
                 "Hyperf\\Retry\\ConfigProvider",
+                "Hyperf\\Resource\\ConfigProvider",
                 "Hyperf\\RpcClient\\ConfigProvider",
                 "Hyperf\\RpcServer\\ConfigProvider",
                 "Hyperf\\Server\\ConfigProvider",
