+++ src/resource/composer.json
@@ -1,0 +1,64 @@
+{
+    "name": "hyperf/resource",
+    "description": "A api resource component for hyperf.",
+    "license": "MIT",
+    "keywords": [
+        "php",
+        "hyperf",
+        "resource"
+    ],
+    "homepage": "https://hyperf.io",
+    "support": {
+        "docs": "https://hyperf.wiki",
+        "issues": "https://github.com/hyperf/hyperf/issues",
+        "pull-request": "https://github.com/hyperf/hyperf/pulls",
+        "source": "https://github.com/hyperf/hyperf"
+    },
+    "require": {
+        "php": ">=7.2",
+        "hyperf/database": "~1.1.0",
+        "hyperf/utils": "~1.1.0",
+        "hyperf/paginator": "~1.1.0",
+        "hyperf/http-server": "~1.1.0",
+        "psr/container": "^1.0",
+        "nesbot/carbon": "^2.0",
+        "psr/event-dispatcher": "^1.0"
+    },
+    "require-dev": {
+        "malukenho/docheader": "^0.1.6",
+        "mockery/mockery": "^1.0",
+        "phpunit/phpunit": "^7.0.0",
+        "friendsofphp/php-cs-fixer": "^2.9"
+    },
+    "suggest": {
+        "hyperf/contract": "Required to use annotations.",
+        "hyperf/di": "Required to use annotations."
+    },
+    "autoload": {
+        "psr-4": {
+            "Hyperf\\Resource\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "HyperfTest\\Resource\\": "tests/"
+        }
+    },
+    "config": {
+        "sort-packages": true
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.1-dev"
+        },
+        "hyperf": {
+            "config": "Hyperf\\Resource\\ConfigProvider"
+        }
+    },
+    "bin": [
+    ],
+    "scripts": {
+        "cs-fix": "php-cs-fixer fix $1",
+        "test": "phpunit --colors=always"
+    }
+}
