+++ src/resource/src/Json/AnonymousResourceCollection.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
-namespace Hyperf\Database\Model\Resource\Json;
+namespace Hyperf\Resource\Json;
 
 class AnonymousResourceCollection extends ResourceCollection
 {
