+++ src/resource/src/Json/PaginatedResourceResponse.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
-namespace Hyperf\Database\Model\Resource\Json;
+namespace Hyperf\Resource\Json;
 
 use Hyperf\Utils\Arr;
